Clarify guide 7.5.1. "Updating the Guide", esp GitHub tasks
This adds and rewords throughout the Guide, 7.5.1. "Updating the
Guide".

There were no instructions for what to do on github.com. Added
steps for forking your own copy of the main repo on GitHub.
Added instructions to push your local branch to the forked repo.
Added instructions to create a pull request. This has two cases:
if GitHub notices the new branch and offers to make a PR, or if
GitHub doesn't notice, and you have to make the PR manually.
Added instructions for participating in the PR discussion, and for
making updates to the PR.

There were no instructions for the Git tasks on the local cloned
repo as part of proposing a change. Added instructions to
checkout master, do git pull to pick up changes from MacPorts
repo, and to create and checkout a branch. Added guidance on
branch naming. Added instructions for pushing the branch to
GitHub.

Also, copyediting throughout. Added intro paragraphs to the top
section and to the subsections.

The <programlisting>...</programlisting> preserve whitespace, so
either they need wonky indentation, or comments to give indentation
without whitespace. This confused me. I added comments throughout,
with notes to help the next editor figure out what's going on:

<!-- [use comments to make programlisting ignore indents] -->