Update adoc's list of ports required to build documentation
In the adoc sources of the Guide, guide/adoc/project.adoc,
7.5. "Updating Documentation",  7.5.1.1 "Preparing Changes",
there is a step to install the required ports, to build the
Guide from XML Docbook sources. Update this list of ports
based on README.md in root.

This matches the corresponding change to guide/xml/project.xml
in commit 36ca5d83a2225c415a17d6420b2d9dcf1754a6fe . It is a
minimal-scope one-line change.

This section of the Guide does not describe how to build the
Guide from adoc sources. Adding such a description is out of
scope for this limited patch.